chore: settingView -> myView 명칭 변경
